socket: set ownership of message queues on creation
Linux creates message queues with the fsuid/fsgid of the calling
process.  Since creation was moved to a separate process in the
parent commit, also move UID/GID handling there.  This drops the
need for allowing setattr in the SELinux policy, and there is no
window where the queue will have unintended ownership.